dom/tests/mochitest/webcomponents/test_custom_element_upgrade_chrome.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1299363
-->
<head>
  <title>Test upgrade steps for custom elements.</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1299363">Bug 1299363</a>
<script type="text/javascript">
SimpleTest.waitForExplicitFinish();

var promises = [];
function test_with_new_window(f, msg) {
  promises.push(new Promise((aResolve) => {
    let iframe = document.createElement('iframe');
    iframe.setAttribute('type', 'content');
    iframe.setAttribute('src', 'http://example.org/chrome/dom/tests/mochitest/webcomponents/test_upgrade_page.html');
    iframe.onload = function() {
      try {
        // Use window from iframe to isolate the test.
        f(iframe.contentWindow, msg);
      } catch (e) {
        SimpleTest.ok(false, e);
      }
      aResolve();
    };
    document.body.appendChild(iframe);
  }));
}
</script>
<!-- Test cases for autonomous element -->
<script type="text/javascript" src="upgrade_tests.js"></script>
<script>
Promise.all(promises).then(() => {
  SimpleTest.finish();
});
</script>
</body>
</html>
